@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
     <AWSProjectType>Lambda</AWSProjectType>
     <ApplicationIcon />
@@ -10,8 +10,9 @@
     <AssemblyName>EntryPoint</AssemblyName>
   </PropertyGroup>
   <ItemGroup>
-	<PackageReference Include="Amazon.Lambda.AspNetCoreServer" Version="2.1.0" />
-	<PackageReference Include="Microsoft.AspNetCore.App" Version="2.1.5" />
+	<PackageReference Include="Amazon.Lambda.AspNetCoreServer" Version="5.0.0" />
+    <PackageReference Include="Amazon.Lambda.RuntimeSupport" Version="1.1.0" />
+    <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="1.0.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\Domain\Domain.csproj" />
